rename and modify migrations

+++ migrations/versions/f8e18fbe2d41_create_tickets_table.py
@@ -1,9 +1,7 @@
 """create tickets table
-
 Revision ID: f8e18fbe2d41
 Revises: 79d78e481ff5
 Create Date: 2017-08-01 19:23:29.919062
-
 """
 from alembic import op
 import sqlalchemy as sa
@@ -21,7 +19,7 @@ def upgrade():
         'tickets',
         sa.Column('id', sa.Integer, primary_key=True),
         sa.Column('ticket_type', sa.String, unique=True),
-        sa.Column('price', sa.Integer),
+        sa.Column('price', sa.Float),
         sa.Column('information', sa.String),
         sa.Column('used', sa.Boolean),
         sa.Column('created_at', sa.DateTime),
